+++ faddr/settings.py
@@ -1,6 +1,6 @@
 """Init default configuration and read configuration from file."""
 
-import pathlib
+from pathlib import Path
 from typing import Any, Dict, List, Tuple
 
 import yaml
@@ -28,7 +28,7 @@ def load_settings(settings_file=None):
 def yaml_config_settings_source(settings: BaseSettings) -> Dict[str, Any]:
     """A simple settings source that loads variables from a YAML file."""
 
-    settings_file_path = pathlib.Path(settings.__config__.settings_file)
+    settings_file_path = Path(settings.__config__.settings_file)
     if settings_file_path.exists():
         try:
             with open(
@@ -76,7 +76,7 @@ class FaddrSettings(BaseSettings):
     """Faddr settings root."""
 
     debug: bool = False
-    templates_dir: pathlib.Path = pathlib.Path(__file__).parent.joinpath("templates")
+    templates_dir: Path = Path(__file__).parent.joinpath("templates")
     database: DatabaseSettings = {}
     rancid: RancidSettings = {}