+++ tests/conftest.py
@@ -1,6 +1,6 @@
 """PyTest fixtures that are accessible by all tests."""
 
-import pathlib
+from pathlib import Path
 
 import pytest
 
@@ -10,9 +10,7 @@ def settings_default():
     """Default settings."""
     settings = {
         "debug": False,
-        "templates_dir": pathlib.Path(__file__)
-        .parent.with_name("faddr")
-        .joinpath("templates"),
+        "templates_dir": Path(__file__).parent.with_name("faddr").joinpath("templates"),
         "database": {"dir": "/var/db/faddr/", "file": "faddr-db.json"},
         "rancid": {
             "dirs": [{"path": "/var/lib/rancid/", "kind": "dir", "mapping": {}}],
@@ -32,9 +30,7 @@ def settings_posix_tmp():
     """Settings with output dir as '/tmp'"""
     settings = {
         "debug": False,
-        "templates_dir": pathlib.Path(__file__)
-        .parent.with_name("faddr")
-        .joinpath("templates"),
+        "templates_dir": Path(__file__).parent.with_name("faddr").joinpath("templates"),
         "database": {"dir": "/tmp/", "file": "faddr-db.json"},
         "rancid": {
             "dirs": [
@@ -58,22 +54,22 @@ def settings_posix_tmp():
 @pytest.fixture
 def settings_file_absent():
     """Absent settings file."""
-    return pathlib.Path("tests/fixtures/nonexistant_settings_file.yaml")
+    return Path("tests/fixtures/nonexistant_settings_file.yaml")
 
 
 @pytest.fixture
 def settings_file_posix_tmp():
     """Settings file with output dir as '/tmp'."""
-    return pathlib.Path("tests/fixtures/faddr_posix_tmp.yaml")
+    return Path("tests/fixtures/faddr_posix_tmp.yaml")
 
 
 @pytest.fixture
 def settings_file_malformed():
     """Corrupted settings file.'"""
-    return pathlib.Path("tests/fixtures/faddr_malformed.yaml")
+    return Path("tests/fixtures/faddr_malformed.yaml")
 
 
 @pytest.fixture
 def settings_file_wrong_format():
     """Wrong format settings file.'"""
-    return pathlib.Path("tests/fixtures/faddr_wrong_format.yml")
+    return Path("tests/fixtures/faddr_wrong_format.yml")