+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- ttp package for configuration files parsing
+- Parser class to work with ttp
+
 ## [0.0.3] - 2022-02-01
 
 ### Added